1. Renamed OpenXml to PowerTools for powertool specific classes to give a clear separation between projects 2. Updated AddSource in DocumentBuilder to handle optional params. This removes need for SourceBuilder 3. Removed Source builder and adjusted examples and code accordingly 4. Refractored more code into appropriate modules
